Use search params for routes
