+++ app/layout.js
@@ -10,8 +10,7 @@ const inter = Inter({
 });
 
 export const metadata = {
-  title: 'Next.js',
-  description: 'Generated by Next.js'
+  title: 'Cocktails'
 };
 
 export default function RootLayout({ children }) {
