+++ styles/global.css
@@ -25,5 +25,7 @@ body {
 }
 
 .container {
+  max-width: 1200px;
+  margin: 0 auto;
   padding: 16px;
 }
